+++ tests/testthat/test-show_orgs.R
@@ -1,6 +1,0 @@
-test_that("show_orgs() returns orgs", {
-  expect_equal(
-    show_orgs(test_gitstats),
-    c("github_test_org", "gitlab_test_group")
-  )
-})